backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0

# Admin interface
sqladmin==0.18.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
